% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{add_column_if_missing}
\alias{add_column_if_missing}
\title{Add Column With Default Value if Not in Data Frame}
\usage{
add_column_if_missing(df, column, default, force = FALSE)
}
\description{
Add Column With Default Value if Not in Data Frame
}
\keyword{internal}
