% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{add_columns_if_missing}
\alias{add_columns_if_missing}
\title{Add Columns With Default Values if Not in Data Frame}
\usage{
add_columns_if_missing(df, defaults, force = FALSE)
}
\description{
Add Columns With Default Values if Not in Data Frame
}
\keyword{internal}
