% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{assign_parameters.junction}
\alias{assign_parameters.junction}
\title{conversion helper}
\usage{
\method{assign_parameters}{junction}(x, infiltration = NULL,
  subcatchment = NULL, subcatchment_typologies = NULL,
  conduit_material = NULL, junction_parameters)
}
\description{
conversion helper
}
\keyword{internal}
