% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{assign_parameters.junctions}
\alias{assign_parameters.junctions}
\title{conversion helper}
\usage{
\method{assign_parameters}{junctions}(
  x,
  infiltration = NULL,
  subcatchment = NULL,
  subcatchment_typologies = NULL,
  conduit_material = NULL,
  junction_parameters
)
}
\description{
conversion helper
}
\keyword{internal}
