% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{assign_parameters.storages}
\alias{assign_parameters.storages}
\title{conversion helper}
\usage{
\method{assign_parameters}{storages}(x, infiltration = NULL,
  subcatchment = NULL, subcatchment_typologies = NULL,
  conduit_material = NULL, junction_parameters = NULL)
}
\description{
conversion helper
}
\keyword{internal}
