% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{assign_parameters.subareas}
\alias{assign_parameters.subareas}
\title{conversion helper}
\usage{
\method{assign_parameters}{subareas}(x, ...)
}
\description{
conversion helper
}
\keyword{internal}
