% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inp.R
\name{build_inp}
\alias{build_inp}
\title{build_inp}
\usage{
build_inp(tmp, inp = NULL, parameter, values, factor)
}
\arguments{
\item{tmp}{Name and path to a template file.}

\item{inp}{Name and path to a new input file.}

\item{parameter}{Names of parameters to be modified.}

\item{values}{The new parameter values.}

\item{factor}{Is the new parameter value a factor?}
}
\description{
build_inp
}
\details{
Builds an SWMM inp file based on a template.
}

