% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{check_package_and_class}
\alias{check_package_and_class}
\title{Helper function}
\usage{
check_package_and_class(x, package = "sf", class = "inp")
}
\description{
Helper function
}
\keyword{internal}
