% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{create_sf_of_pt}
\alias{create_sf_of_pt}
\title{helper function to create a simple feature object from point geometry
function is not exported}
\usage{
create_sf_of_pt(x)
}
\arguments{
\item{x}{a tibble to be converted}
}
\description{
helper function to create a simple feature object from point geometry
function is not exported
}
\keyword{internal}
