% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_exec_on_linux_or_mac}
\alias{.get_exec_on_linux_or_mac}
\title{Determine the latest swmm5.exe in the specified  program files folder}
\usage{
.get_exec_on_linux_or_mac(path, exe)
}
\description{
Determine the latest swmm5.exe in the specified  program files folder
}
\keyword{internal}
