% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_out_helper.R
\name{.get_iIndex}
\alias{.get_iIndex}
\title{Get the names of SWMM elements}
\usage{
.get_iIndex(list_of_elements, iType = NULL, object_name = NULL)
}
\description{
Get the names of SWMM elements
}
\keyword{internal}
