% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swmm.R
\name{.get_latest_swmm_windows}
\alias{.get_latest_swmm_windows}
\title{Determine the latest swmm5.exe in the specified  program files folder}
\usage{
.get_latest_swmm_windows(
  path = "c:/Program Files (x86)/",
  prefix = "EPA SWMM",
  exe = "swmm5.exe"
)
}
\description{
Determine the latest swmm5.exe in the specified  program files folder
}
\keyword{internal}
