% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swmm.R
\name{.get_path_to_exec}
\alias{.get_path_to_exec}
\title{Check if SWMM executable is in one of the expected directories.}
\usage{
.get_path_to_exec()
}
\description{
Check if SWMM executable is in one of the expected directories.
}
\keyword{internal}
