% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_out.R
\name{get_out_content}
\alias{get_out_content}
\title{Get the content of an .out file.}
\usage{
get_out_content(file = "")
}
\arguments{
\item{file}{The file to be read.}
}
\value{
A list showing the available content.
}
\description{
This function opens an .out file and lists all available time series data.
Currently, the list is returned 'as is' which might change in future.
It belongs to a set of helper functions which aim to simplify
the work with .out files. The lifecycle of this function is considered
experimental.
}
\examples{
\dontrun{
content <- get_out_content("model.out")
} 
}
