% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp_to_files.R
\name{inp_to_files}
\alias{inp_to_files}
\title{Convert SWMM's .inp to .shp and txt files}
\usage{
inp_to_files(x, name, path_out)
}
\arguments{
\item{x}{An object of class inp.}

\item{name}{Give a name for the current model, e.g. "Example1".}

\item{path_out}{Writeable directory name where to save the converted files. Folders: dat, shp and txt will be created if not existent.}
}
\value{
.dat, .shp and/or .txt files.
}
\description{
Convert SWMM's .inp to .shp and txt files
}
