% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_list_of_sections.R
\name{input_to_list_of_sections}
\alias{input_to_list_of_sections}
\title{Reads input data and checks data for completeness, returns a list_of_sections which can be processed further.}
\usage{
input_to_list_of_sections(
  path_options,
  subcatchment,
  subcatchment_typologies,
  junctions,
  junction_parameters,
  outfalls,
  conduits,
  conduit_material,
  path_timeseries,
  infiltration,
  pumps,
  path_pump_curve,
  weirs,
  storage,
  path_storage_curve
)
}
\description{
Reads input data and checks data for completeness, returns a list_of_sections which can be processed further.
}
\keyword{internal}
