% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_sections_to_inp.R
\name{list_of_sections_to_inp}
\alias{list_of_sections_to_inp}
\title{Reads input data which was checked for completeness and converts/processes them to an object of class inp.}
\usage{
list_of_sections_to_inp(
  list_of_sections,
  infiltration,
  subcatchment_typologies,
  conduit_material,
  junction_parameters
)
}
\description{
Reads input data which was checked for completeness and converts/processes them to an object of class inp.
}
\keyword{internal}
