% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{openSWMMOutput}
\alias{openSWMMOutput}
\title{Open a SWMM output file}
\usage{
openSWMMOutput(SWMMoutfile, timezone = "")
}
\arguments{
\item{SWMMoutfile}{character, path to output file}

\item{timezone}{time zone specification of the dates. If \code{""}, the local time zone
is used.}
}
\value{
a \code{SWMMfile} object
}
\description{
opens a binay SWMM output file
}
\author{
Andreas Scheidegger
}
