% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.analysis_options}
\alias{parse_section.analysis_options}
\title{import helper}
\usage{
\method{parse_section}{analysis_options}(x, ...)
}
\description{
import helper
}
\keyword{internal}
