% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.conduit_surcharge_summary}
\alias{parse_section.conduit_surcharge_summary}
\title{import helper}
\usage{
\method{parse_section}{conduit_surcharge_summary}(x, ...)
}
\description{
import helper
}
\keyword{internal}
