% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.highest_continuity_errors}
\alias{parse_section.highest_continuity_errors}
\title{import helper}
\usage{
\method{parse_section}{highest_continuity_errors}(x, ...)
}
\description{
import helper
}
\keyword{internal}
