% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.highest_flow_instability_indexes}
\alias{parse_section.highest_flow_instability_indexes}
\title{import helper}
\usage{
\method{parse_section}{highest_flow_instability_indexes}(x, ...)
}
\description{
import helper
}
\keyword{internal}
