% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.link_flow_summary}
\alias{parse_section.link_flow_summary}
\title{import helper}
\usage{
\method{parse_section}{link_flow_summary}(x, ...)
}
\description{
import helper
}
\keyword{internal}
