% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.pumping_summary}
\alias{parse_section.pumping_summary}
\title{import helper}
\usage{
\method{parse_section}{pumping_summary}(x, ...)

\method{parse_section}{pumping_summary}(x, ...)
}
\description{
import helper

import helper
}
\keyword{internal}
