% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.runoff_quantity_continuity}
\alias{parse_section.runoff_quantity_continuity}
\title{import helper}
\usage{
\method{parse_section}{runoff_quantity_continuity}(x, ...)
}
\description{
import helper
}
\keyword{internal}
