% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{parse_section.time_step_critical_elements}
\alias{parse_section.time_step_critical_elements}
\title{import helper}
\usage{
\method{parse_section}{time_step_critical_elements}(x, ...)
}
\description{
import helper
}
\keyword{internal}
