% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inp.R
\name{plot.inp}
\alias{plot.inp}
\title{Plot a swmm model structure}
\usage{
\method{plot}{inp}(x, ...)
}
\arguments{
\item{x}{An inp object}

\item{...}{currently ignored}
}
\description{
This function reads an object of class 'inp' and extracts
subcatchments, links, junctions and raingages. All objects are converted to simple feautures
geometries and passed to \code{\link[ggplot2]{geom_sf}}.
}
\examples{
 
\dontrun{
inp <- read_inp("model.inp")
plot(inp)
}
}
