% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{print.SWMMfile}
\alias{print.SWMMfile}
\title{Print main properties of a SWMM output file}
\usage{
\method{print}{SWMMfile}(SWMMfile)
}
\arguments{
\item{SWMMfile}{a \code{SWMMfile} object}
}
\value{
nothing
}
\author{
Andreas Scheidegger
}
