% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inp.R
\name{read_inp}
\alias{read_inp}
\title{read_inp}
\usage{
read_inp(inp)
}
\arguments{
\item{inp}{Name and path to an input file.}
}
\description{
Read the SWMM .inp file
}
\details{
Reads an SWMM .inp file and creates a list with corresponding SWMM-sections.
This function reads the .inp in a very simplified but reasonable manner.
It's main purpose is to glance the model structure and options set.
}

