% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inp.R
\name{read_inp2}
\alias{read_inp2}
\title{Read SWMM's .inp file}
\usage{
read_inp2(x, rm.comment = TRUE)
}
\arguments{
\item{rm.comment}{Should lines with comments starting with a ";" be discarded?}

\item{inp}{Name (incl. path) to an input file.}
}
\value{
An object of class \code{inp}
}
\description{
Reads a SWMM .inp file and creates a list with corresponding SWMM sections.
}
\examples{
 
\dontrun{
inp_sections <- read_inp("model.inp")
} 
}
