% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lid_rpt.R
\name{read_lid_rpt}
\alias{read_lid_rpt}
\title{Read SWMM's LID Report File}
\usage{
read_lid_rpt(x, ...)
}
\arguments{
\item{x}{Name (incl. path) to a LID report file.}

\item{...}{optional arguments passed to \code{\link[readr]{read_table2}}}
}
\value{
A tibble
}
\description{
Reads a SWMM's LID Report File and returns a tibble
}
\examples{
 
\dontrun{
tbl_lid_rpt <- read_lid_rpt("lid_rpt.txt")
} 
}
