% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rpt.R
\name{read_rpt}
\alias{read_rpt}
\title{Read SWMM's .rpt file}
\usage{
read_rpt(x)
}
\arguments{
\item{x}{Name (incl. path) to an report file.}
}
\value{
An object of class \code{rpt}
}
\description{
Reads a SWMM .rpt file and creates a list with corresponding results sections.
}
\examples{
 
\dontrun{
list_of_rpt_results <- read_rpt("model.rpt")
} 
}
