% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{readLinks}
\alias{readLinks}
\title{read SWM results for links}
\usage{
readLinks(SWMMfile, names, variables = NULL)
}
\arguments{
\item{SWMMfile}{a \code{SWMMfile} object}

\item{names}{vector of names of the links to be read. Partial matching is supported.}

\item{variables}{vector of variables to be read for each link, see Details below. If \code{NULL}, all variables are read. Partial matching is supported.}
}
\value{
A named list of separate \code{xts} objects for each variable.
The \code{xts} objects contain the data for all links.
}
\description{
Read link results from a SWMM file.
}
\details{
The following variables are available for links:
\itemize{
\item{"flow rate"}{}
\item{"flow depth"}{}
\item{"flow velocity"}{}
\item{"Froude number"}{}
\item{"capacity"}{fraction of conduit filled}
\item{"PollutantName concentration"} {Only if pollutions are modeled.}
}
}
\author{
Andreas Scheidegger
}
