% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{readNodes}
\alias{readNodes}
\title{read SWM results for nodes}
\usage{
readNodes(SWMMfile, names = NULL, variables = NULL)
}
\arguments{
\item{SWMMfile}{a \code{SWMMfile} object}

\item{names}{vector of names of the nodes to be read. If \code{NULL}, all nodes are read. Partial matching is supported.}

\item{variables}{vector of variables to be read for each node, see Details below. If \code{NULL}, all variables are read. Partial matching is supported.}
}
\value{
A named list of separate \code{xts} objects for each variable.
The \code{xts} objects contain the data for all nodes.
}
\description{
Read node results from a SWMM file.
}
\details{
The following variables are available for nodes:
\itemize{
\item{"depth of water above invert"}{}
\item{"hydraulic head"}{}
\item{"volume of stored + ponded water"}{}
\item{"lateral inflow"}{}
\item{"total inflow"}{lateral + upstream}
\item{"flow lost to flooding"}{}
\item{"PollutantName concentration"} {Only if pollutions are modeled.}
}
}
\author{
Andreas Scheidegger
}
