% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{readSingleResult}
\alias{readSingleResult}
\title{Read a single SWMM result}
\usage{
readSingleResult(SWMMfile, iType, iIndex, vIndex, period)
}
\arguments{
\item{SWMMfile}{a \code{SWMMfile} object}

\item{iType}{\code{0}, \code{1}, \code{2}, or \code{3}. See Detail below.}

\item{iIndex}{integer, the position of the subcatch, subcatch, node, link or sys object}

\item{vIndex}{integer, index of the variable. See Detail below.}

\item{period}{integer, index of the period (time step)}
}
\value{
a scalar
}
\description{
This function used internally only. Reads a single result from a SWMM outputfile.
}
\details{
\code{iType} should be 0 for Subcatchments, 1 for nodes, 2 for links, or 3 for system variables.

\code{vIndex} should be selected from this lists below, depending on whether \code{iType} is a subcatchment, a link, or a node.
                       
Number of subcatchment variables (currently 6 + number of pollutants).
Code number of each subcatchment variable:
0 for rainfall (in/hr or mm/hr),
1 for snow depth (in or mm),
2 for evaporation + infiltration losses (in/hr or mm/hr),
3 for runoff rate (flow units),
4 for groundwater outflow rate (flow units),
5 for groundwater water table elevation (ft or m),
6 for runoff concentration of first pollutant,
5 + N for runoff concentration of N-th pollutant.

Number of node variables (currently 6 + number of pollutants)
Code number of each node variable:
0 for depth of water above invert (ft or m),
1 for hydraulic head (ft or m),
2 for volume of stored + ponded water (ft3 or m3),
3 for lateral inflow (flow units),
4 for total inflow (lateral + upstream) (flow units),
5 for flow lost to flooding (flow units),
6 for concentration of first pollutant,
5 + N for concentration of N-th pollutant.

Number of link variables (currently 5 + number of pollutants)
Code number of each link variable:
                           
0 for flow rate (flow units),
1 for flow depth (ft or m),
2 for flow velocity (ft/s or m/s),
3 for Froude number,
4 for capacity (fraction of conduit filled),
5 for concentration of first pollutant,
4 + N for concentration of N-th pollutant.

Number of system-wide variables (currently 14)
Code number of each system-wide variable:
0 for air temperature (deg. F or deg. C),
1 for rainfall (in/hr or mm/hr),
2 for snow depth (in or mm),
3 for evaporation + infiltration loss rate (in/hr or mm/hr),
4 for runoff flow (flow units),
5 for dry weather inflow (flow units),
6 for groundwater inflow (flow units),
7 for RDII inflow (flow units),
8 for user supplied direct inflow (flow units),
9 for total lateral inflow (sum of variables 4 to 8) (flow units),
10 for flow lost to flooding (flow units),
11 for flow leaving through outfalls (flow units),
12 for volume of stored water (ft3 or m3),
13 for evaporation rate (in/day or mm/day)
}
\author{
Andreas Scheidegger
}
