% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{readSubcatchments}
\alias{readSubcatchments}
\title{read SWM results for subcatchments}
\usage{
readSubcatchments(SWMMfile, names, variables = NULL)
}
\arguments{
\item{SWMMfile}{a \code{SWMMfile} object}

\item{names}{vector of names of the subcatchment to be read. Partial matching is supported.}

\item{variables}{vector of variables to be read for each subcatchment, see Details below. If \code{NULL}, all variables are read. Partial matching is supported.}
}
\value{
A named list of separate \code{xts} objects for each variable.
The \code{xts} objects contain the data for all subcatchment.
}
\description{
Read subcatchment results from a SWMM file.
}
\details{
The following variables are available for subcatchments:
\itemize{
\item{"rainfall"}{}
\item{"snow depth"}{}
\item{"evaporation losses"}{}
\item{"infiltration losses"}{}
\item{"runoff"}{}
\item{"groundwater outflow"}{}
\item{"groundwater water table elevation"}{}
\item{"unsaturated zone moisture content"}{fraction}
\item{"PollutantName"} {Only if pollutions are modeled.}
}
}
\author{
Andreas Scheidegger
}
