% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{readSystem}
\alias{readSystem}
\title{read SWM results for the system}
\usage{
readSystem(SWMMfile, variables = NULL)
}
\arguments{
\item{SWMMfile}{a \code{SWMMfile} object}

\item{variables}{vector of variables to be read, see Details below. If \code{NULL}, all variables are read.
Partial matching is supported.}
}
\value{
A named list of \code{xts} objects for each variable.
}
\description{
Reads the system results from a SWMM file.
}
\details{
The following variables are available for the system:
\itemize{
\item{"air temperature"}{}
\item{"rainfall"}{}
\item{"snow depth"}{}
\item{"evaporation + infiltration loss rate"}{}
\item{"runoff flow"}{}
\item{"dry weather inflow"}{}
\item{"groundwater inflow"}{}
\item{"RDII inflow"}{}
\item{"direct inflow"}{user supplied}
\item{"total lateral inflow"}{sum of runoff flow plus all inflows}
\item{"flow lost to flooding"}{}
\item{"flow leaving through outfalls"}{}
\item{"volume of stored water"}{}
\item{"actual evaporation rate"}{}
\item{"potential evaporation rate"}
}
}
\author{
Andreas Scheidegger
}
