% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swmm.R
\name{run_swmm}
\alias{run_swmm}
\title{run_swmm}
\usage{
run_swmm(inp, rpt = NULL, out = NULL,
  exe = "C:/Program Files (x86)/EPA SWMM 5.1/swmm5.exe", show.progress = T)
}
\arguments{
\item{inp}{Name and path to an input file.}

\item{rpt}{Name and path to an report file.}

\item{out}{Name and path to an out file.}

\item{exe}{Name and path to swmm5.exe.}

\item{show.progress}{logical. Defines whether simulation progress is printed.}
}
\description{
run_swmm
}
\details{
Run a SWMM inp file.
}

