% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SWMM.r
\name{runSWMM}
\alias{runSWMM}
\title{Run a SWMM simulation with given parameters}
\usage{
runSWMM(parameters = NULL, inputfile, reportfilename = "report.rpt",
  outputfilename = "output.out", SWMMexe = "swmm5.exe", quite = FALSE)
}
\arguments{
\item{parameters}{\emph{named} vector of parameters or \code{NULL}}

\item{inputfile}{path/name to input file. If \code{parameters} is not \code{NULL}
the file is used as template, see Details below.}

\item{reportfilename}{path/name of the report file}

\item{outputfilename}{path/name of output file}

\item{SWMMexe}{path to SWMM executable}

\item{quite}{Boolean. If \code{FALSE} consol output is shown.}
}
\value{
nothing
}
\description{
Run SWMM simulations. If parameters are given the input file is modified accordingly.
}
\details{
If the argument \code{parameters} is \code{NULL}, the inut file is used as is.
If \code{parameters} is a named vector the input file is treated as template, i.e. all
\code{\{parametername\}} are replaced with the value of \code{parameters}.
}
\author{
Andreas Scheidegger
}
