% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{separate_into}
\alias{separate_into}
\title{helper function using defaults for arguments to tidyr::separate()}
\usage{
separate_into(
  x,
  into,
  fill = "left",
  sep = "\\\\s+",
  extra = "merge",
  col = "value",
  convert = TRUE,
  ...
)
}
\description{
helper function using defaults for arguments to tidyr::separate()
}
\keyword{internal}
