% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_section.R
\name{skip_head}
\alias{skip_head}
\title{helper function skippting the first n rows of a data frame}
\usage{
skip_head(df, n)
}
\description{
helper function skippting the first n rows of a data frame
}
\keyword{internal}
