% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.inp}
\alias{summary.inp}
\title{Show summary of a swmm model structure}
\usage{
\method{summary}{inp}(object, ...)
}
\arguments{
\item{object}{An object of class 'inp', created by \code{\link{read_inp}}.}

\item{...}{Additional parameters (currently ignored).}
}
\value{
The input is returned invisibly.
}
\description{
Show summary of a swmm model structure
}
\examples{
\dontrun{
x <- read_inp("model.inp")
summary(x)
} 
}
