\name{swmmr-package}
\alias{swmmr-package}
\alias{swmmr}
\docType{package}
\title{
\packageTitle{swmmr}
}
\description{
\packageDescription{swmmr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{swmmr}
\packageIndices{swmmr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{swmmr}

Maintainer: \packageMaintainer{swmmr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
