% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\docType{package}
\name{SWMMR}
\alias{SWMMR}
\alias{SWMMR-package}
\title{SWMMR - R interface to the Storm Water Management Model (SWMM)}
\description{
This package enables running a SWMM simulation, modifying the model inputs,
and reading the binary files.
}
\details{
The code is based on the work of Peter Steinberg \url{https://github.com/PeterDSteinberg/RSWMM}.

An alternative is the \code{swmmr} package by Dominik Leutnant \url{https://github.com/dleutnant/swmmr}. Output files
read in C++ which should be faster.
}
\author{
Andreas Scheidegger
}
