% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{assign_parameters.conduits}
\alias{assign_parameters.conduits}
\title{conversion helper}
\usage{
\method{assign_parameters}{conduits}(x, infiltration = NULL,
  subcatchment = NULL, subcatchment_typologies = NULL, conduit_material,
  junction_parameters = NULL)
}
\description{
conversion helper
}
\keyword{internal}
