% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_parameters.R
\name{assign_parameters.infiltration}
\alias{assign_parameters.infiltration}
\title{conversion helper}
\usage{
\method{assign_parameters}{infiltration}(x, infiltration, subcatchment = NULL,
  subcatchment_typologies = NULL, conduit_material = NULL,
  junction_parameters = NULL)
}
\description{
conversion helper
}
\keyword{internal}
