% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SWMM.r
\name{closeSWMMOutput}
\alias{closeSWMMOutput}
\title{Closes a SWMM file handler}
\usage{
closeSWMMOutput(SWMMfile)
}
\arguments{
\item{SWMMfile}{a SWMM file handler}
}
\value{
nothing
}
\author{
Andreas Scheidegger
}
