% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{convert_to_sf}
\alias{convert_to_sf}
\alias{junctions_to_sf}
\alias{subcatchments_to_sf}
\alias{links_to_sf}
\alias{inp_to_sf}
\title{Convert swmm objects to simple feauture geometries}
\usage{
junctions_to_sf(x)

subcatchments_to_sf(x)

links_to_sf(x)

inp_to_sf(x)
}
\arguments{
\item{x}{An object of class 'swmm_inp', created by 'read_inp'}
}
\value{
A simple feature
}
\description{
\itemize{
\item \code{junctions_to_sf()}: converts junctions to simple features
\item \code{links_to_sf()}: converts links to simple features
\item \code{subcatchments_to_sf()}: converts subcatchments to simple features
\item \code{inp_to_sf()}: converts junctions, links and subcatchments to simple features
}
}
\seealso{
\code{\link[ggplot2]{geom_sf}}
}
