% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_out.R
\name{get_out_version}
\alias{get_out_version}
\title{Get the swmm version the .out file was generated with}
\usage{
get_out_version(file = "")
}
\arguments{
\item{file}{The file to be read.}
}
\value{
A vector of type integer
}
\description{
This function opens an .out file and extract the swmm version the file was
generated with. It belongs to a set of helper functions which aim to simplify
the work with .out files. The lifecycle of this function is considered
experimental.
}
\examples{
\dontrun{
version <- get_out_version("model.out")
} 
}
