% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_list_of_sections.R
\name{input_to_list_of_sections}
\alias{input_to_list_of_sections}
\title{Reads input data and checks data for completeness, returns a list_of_sections which can be processed further.}
\usage{
input_to_list_of_sections(
  path_options,
  subcatchment = NULL,
  subcatchment_typologies = NULL,
  junctions = NULL,
  junction_parameters = NULL,
  outfalls = NULL,
  conduits = NULL,
  conduit_material = NULL,
  path_timeseries = NULL,
  infiltration = NULL,
  pumps = NULL,
  path_pump_curve = NULL,
  weirs = NULL,
  storage = NULL,
  path_storage_curve = NULL
)
}
\description{
Reads input data and checks data for completeness, returns a list_of_sections which can be processed further.
}
\keyword{internal}
