% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inp.R
\name{plot.swmm_inp}
\alias{plot.swmm_inp}
\title{Plot a swmm model structure}
\usage{
\method{plot}{swmm_inp}(x, ...)
}
\arguments{
\item{x}{An swmm_inp object}

\item{...}{currently ignored}
}
\description{
This function reads an object of class 'swmm_inp' and extracts
subcatchments, links and junctions. All objects are converted to simple feautures
geometries and passed to \code{\link[ggplot2]{geom_sf}}.
}
\examples{
 
\dontrun{
swmm_inp <- read_inp("model.inp")
plot(swmm_inp)
}
}
