% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.inp}
\alias{autoplot.inp}
\title{Plot a swmm model structure using ggplot2}
\usage{
autoplot.inp(x, ...)
}
\arguments{
\item{x}{An inp object}

\item{...}{currently ignored}
}
\description{
This function reads an object of class 'inp'. All objects are
converted to simple feature geometries via \code{\link[swmmr]{inp_to_sf}} and
finally passed to \code{\link[ggplot2]{geom_sf}}. It allows to quickly
visualize a model structure.
}
\note{
Lifecycle: experimental
}
\examples{
\dontrun{
inp <- read_inp("model.inp")
autoplot(inp)
}
}
