% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inp.R
\name{read_inp}
\alias{read_inp}
\title{Read SWMM's .inp file}
\usage{
read_inp(inp, rm.comment = TRUE)
}
\arguments{
\item{inp}{Name (incl. path) to an input file.}

\item{rm.comment}{Should lines with comments starting with a ";" be discarded?}
}
\value{
A list with SWMM inp sections.
}
\description{
Reads a SWMM .inp file and creates a list with corresponding SWMM sections.
This function reads the .inp in a very simplified but reasonable manner.
It's main purpose is to glance the model structure and simulation options set.
}
\examples{
 
\dontrun{
inp_sections <- read_inp("model.inp")
} 
}
