% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inp.R
\name{read_inp}
\alias{read_inp}
\title{Read SWMM's .inp file}
\usage{
read_inp(x, rm.comment = TRUE, ...)
}
\arguments{
\item{x}{Name (incl. path) to an input file.}

\item{rm.comment}{Should lines with comments starting with a ";" be discarded?}

\item{...}{optional arguments passed to \code{\link[readr]{read_lines}}.}
}
\value{
An object of class \code{inp}
}
\description{
Reads a SWMM .inp file and creates a list with corresponding SWMM sections.
}
\examples{
 
\dontrun{
list_of_inp_sections <- read_inp("model.inp")
} 
}
